backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
google-generativeai
pydantic
pydantic-settings
sqlalchemy
psycopg2-binary
python-dotenv
